requirements.txt
boto3==1.34.144
matplotlib==3.7.5
numpy==1.24.3
pillow==10.3.0
seaborn==0.13.2
scikit-learn==1.3.2
tensorflow_probability==0.19.0
# (note tensorflow itself installed separately for cpu or gpu version)
